// src/utils/enums.ts
export enum ColumnType {
  BACKLOG = 'Backlog  ⛔',
  DESIGN = 'Design 🎨',
  TO_DO = 'To Do  ⏰',
  IN_PROGRESS = 'In Progress  🧑🏻‍💻',
  CODE_REVIEW = 'Code Review  📃',
  TESTING = 'Testing  🪲',
  DEPLOYED = 'Deployed  🚀',
  MAINTAIN = 'Under Maintainance  ⚙︰'
}

export enum ItemType {
  TASK = 'Task',
}